requirements.txt
streamlit>=1.36.0
python-dotenv>=1.0.1
langchain>=0.2.6
langchain-community>=0.2.6
langchain-core>=0.2.10
langchain-openai>=0.1.0
pinecone>=4.0.0
sentence-transformers>=3.0.1
pypdf>=4.2.0
huggingface-hub>=0.23.4
